requirements.txt
Django==4.0.10
django-environ==0.10.0
djangorestframework==3.14.0
drf-spectacular==0.26.0
whitenoise==6.4.0
gunicorn==20.1.0
dj-database-url==2.2.0
django-rest-swagger==2.2.0
django-cors-headers==3.14.0
psycopg2-binary==2.9.5
pillow==9.4.0